[774] Move several types into core_types module
