+++ bob-backend/src/lib.rs
@@ -22,8 +22,9 @@ pub(crate) mod prelude {
             cluster::Cluster as ClusterConfig,
             node::{BackendType, Node as NodeConfig, Pearl as PearlConfig},
         },
-        data::{BobData, BobKey, BobMeta, DiskPath, VDiskId},
+        data::{BobData, BobKey, BobMeta},
         operation_options::{BobPutOptions, BobGetOptions, BobDeleteOptions},
+        core_types::{DiskPath, VDiskId},
         node::NodeName,
         error::Error,
         mapper::Virtual,
