+++ bob-common/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bob_client;
 pub mod configs;
 mod name_types; // Private module. Inner types should be re-exported where they are needed
+pub mod core_types;
 pub mod data;
 pub mod operation_options;
 pub mod error;
